refactor(service-members): split service member / invitation comps
