+++ packages/app/pages/services/members.js
@@ -1,6 +1,8 @@
+import { BoxWrapper } from "@socialgouv/emjpm-ui-core";
 import React, { useContext } from "react";
 
 import { LayoutServices } from "../../src/components/Layout";
+import { ServiceMemberInvitations } from "../../src/components/ServiceMemberInvitations";
 import { ServiceMembers } from "../../src/components/ServiceMembers";
 import { UserContext } from "../../src/components/UserContext";
 import { withAuthSync } from "../../src/util/auth";
@@ -11,7 +13,10 @@ const ServiceMembersPage = () => {
 
   return (
     <LayoutServices>
-      <ServiceMembers service={service} isAdmin={is_admin} />
+      <BoxWrapper mt={6} px={2}>
+        <ServiceMemberInvitations service={service} isAdmin={is_admin} />
+        <ServiceMembers service={service} isAdmin={is_admin} />
+      </BoxWrapper>
     </LayoutServices>
   );
 };
