+++ packages/app/src/components/ServiceMemberInvitations/index.js
@@ -1,0 +1,3 @@
+import { ServiceMemberInvitations } from "./ServiceMemberInvitations";
+
+export { ServiceMemberInvitations };
