+++ packages/app/src/components/ServiceMemberInvitations/queries.js
@@ -1,20 +1,15 @@
 import gql from "graphql-tag";
 
-export const SERVICE_MEMBERS = gql`
-  query ServiceMembers($serviceId: Int!) {
-    service_members(
+export const SERVICE_MEMBER_INVITATIONS = gql`
+  query ServiceMemberInvitations($serviceId: Int!) {
+    service_member_invitations(
       where: { service_id: { _eq: $serviceId } }
-      order_by: { user: { created_at: desc } }
+      order_by: { created_at: desc }
     ) {
       id
-      is_admin
-      user_id
-      user {
-        id
-        email
-        created_at
-        active
-      }
+      email
+      created_at
+      sent_at
     }
   }
 `;
