+++ packages/app/src/components/ServiceMembers/ServiceMembers.js
@@ -5,9 +5,9 @@ import React from "react";
 import { Box, Flex } from "rebass";
 import { Trash } from "styled-icons/boxicons-regular";
 
+import { ServiceMemberInvitations } from "../ServiceMemberInvitations";
 import { DELETE_SERVICE_MEMBER, UPDATE_SERVICE_MEMBER_IS_ADMIN } from "./mutations";
 import { SERVICE_MEMBERS } from "./queries";
-import { ServiceMemberInvitations } from "./ServiceMemberInvitations";
 import {
   listActionsStyle,
   listActionStyle,
@@ -60,62 +60,52 @@ const ServiceMembers = props => {
   const { service_members } = data;
 
   return (
-    <BoxWrapper mt={6} px={2}>
-      <Box mb={4}>
-        <Heading2 width={[1]} mb="2">
-          Invitations
-        </Heading2>
-        <Card width={[1]} p={4} mb={4}>
-          <ServiceMemberInvitations isAdmin={isAdmin} service={service} />
-        </Card>
-      </Box>
-      <Box mb={4}>
-        <Heading2 width={[1]} mb="2">
-          Membres
-        </Heading2>
-        <Card p={4} width={[1]}>
-          <Heading4 mb="2">Liste des membres ({service_members.length})</Heading4>
-          {service_members.map((member, i) => (
-            <Flex sx={() => listStyle(i)} index={i} key={member.user.email}>
-              <Box sx={listIdStyle}>{member.id}.</Box>
-              <Box sx={listEmailStyle}>{member.user.email}</Box>
-              <Text sx={listDateStyle}>
-                {`Inscrit le `}
-                {format(new Date(member.user.created_at), "dd/MM/yyyy")}
-              </Text>
-              <Box sx={listAdminStyle}>
-                {isAdmin ? (
-                  <Select
-                    id="urgent"
-                    name="urgent"
-                    width={200}
-                    placeholder="Est-ce une demande urgente"
-                    value={member.is_admin ? isAdminOptions[0] : isAdminOptions[1]}
-                    hasError
-                    onChange={({ value }) => handleIsAdminUpdate(member.id, value)}
-                    options={isAdminOptions}
-                  />
-                ) : member.is_admin ? (
-                  "Administrateur"
-                ) : (
-                  "Membre"
-                )}
-              </Box>
-              <Text sx={listDateStyle}>
-                {member.user.active ? "Activé" : "En attente de d'activation"}
-              </Text>
-              {isAdmin && (
-                <Box sx={listActionsStyle}>
-                  <Box sx={listActionStyle} onClick={() => handleDelete(member.id)}>
-                    <Trash title="Supprimer" size="22" />
-                  </Box>
-                </Box>
+    <Box mb={4}>
+      <Heading2 width={[1]} mb="2">
+        Membres
+      </Heading2>
+      <Card p={4} width={[1]}>
+        <Heading4 mb="2">Liste des membres ({service_members.length})</Heading4>
+        {service_members.map((member, i) => (
+          <Flex sx={() => listStyle(i)} index={i} key={member.user.email}>
+            <Box sx={listIdStyle}>{member.id}.</Box>
+            <Box sx={listEmailStyle}>{member.user.email}</Box>
+            <Text sx={listDateStyle}>
+              {`Inscrit le `}
+              {format(new Date(member.user.created_at), "dd/MM/yyyy")}
+            </Text>
+            <Box sx={listAdminStyle}>
+              {isAdmin ? (
+                <Select
+                  id="urgent"
+                  name="urgent"
+                  width={200}
+                  placeholder="Est-ce une demande urgente"
+                  value={member.is_admin ? isAdminOptions[0] : isAdminOptions[1]}
+                  hasError
+                  onChange={({ value }) => handleIsAdminUpdate(member.id, value)}
+                  options={isAdminOptions}
+                />
+              ) : member.is_admin ? (
+                "Administrateur"
+              ) : (
+                "Membre"
               )}
-            </Flex>
-          ))}
-        </Card>
-      </Box>
-    </BoxWrapper>
+            </Box>
+            <Text sx={listDateStyle}>
+              {member.user.active ? "Activé" : "En attente de d'activation"}
+            </Text>
+            {isAdmin && (
+              <Box sx={listActionsStyle}>
+                <Box sx={listActionStyle} onClick={() => handleDelete(member.id)}>
+                  <Trash title="Supprimer" size="22" />
+                </Box>
+              </Box>
+            )}
+          </Flex>
+        ))}
+      </Card>
+    </Box>
   );
 };
